move fake.go to testing package for reusability
