+++ staging/src/k8s.io/apiserver/pkg/endpoints/responsewriter/testing/fake.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package responsewriter
+package testing
 
 import (
 	"bufio"
